Move RecentListens, UserFeedback and UserPageHeading to a folder named `user` and remove fake data from YearInMusic.tsx
